Add BMWConnectedDriveBaseEntity mixin, remove unnecessary type casts
